+++ kcache.h
@@ -49,7 +49,7 @@ extern "C" {
 */
 #ifndef LASVM_KERNEL_T_DEFINED
 #define LASVM_KERNEL_T_DEFINED
-typedef double (*lasvm_kernel_t)(int i, int j, void* closure);
+typedef double (*lasvm_kernel_t)(long i, long j, void* closure);
 #endif
 
 
@@ -98,7 +98,7 @@ long lasvm_kcache_get_current_size(lasvm_kcache_t *self);
    Returns the possibly cached value of the Gram matrix element (<i>,<j>).
    This function will not modify the cache geometry.
  */
-double lasvm_kcache_query(lasvm_kcache_t *self, int i, int j);
+double lasvm_kcache_query(lasvm_kcache_t *self, long i, long j);
 
 /* --- lasvm_kcache_query_row
    Returns the <len> first elements of row <i> of the Gram matrix.
@@ -108,19 +108,19 @@ double lasvm_kcache_query(lasvm_kcache_t *self, int i, int j);
    and vice-versa.
 */
 
-float *lasvm_kcache_query_row(lasvm_kcache_t *self, int i, int len);
+float *lasvm_kcache_query_row(lasvm_kcache_t *self, long i, long len);
 
 /* --- lasvm_kcache_status_row
    Returns the number of cached entries for row i.
 */
 
-int lasvm_kcache_status_row(lasvm_kcache_t *self, int i);
+long lasvm_kcache_status_row(lasvm_kcache_t *self, long i);
 
 /* --- lasvm_kcache_discard_row
    Indicates that we wont need row i in the near future.
 */
 
-void lasvm_kcache_discard_row(lasvm_kcache_t *self, int i);
+void lasvm_kcache_discard_row(lasvm_kcache_t *self, long i);
 
 
 /* --- lasvm_kcache_i2r
@@ -129,8 +129,8 @@ void lasvm_kcache_discard_row(lasvm_kcache_t *self, int i);
    the conversion table from example index to row position and vice-versa. 
 */
 
-int *lasvm_kcache_i2r(lasvm_kcache_t *self, int n);
-int *lasvm_kcache_r2i(lasvm_kcache_t *self, int n);
+long *lasvm_kcache_i2r(lasvm_kcache_t *self, long n);
+long *lasvm_kcache_r2i(lasvm_kcache_t *self, long n);
 
 
 /* --- lasvm_kcache_swap_rr
@@ -141,9 +141,9 @@ int *lasvm_kcache_r2i(lasvm_kcache_t *self, int n);
    or by indicating the example number (<i1>, <i2>).
 */
 
-void lasvm_kcache_swap_rr(lasvm_kcache_t *self, int r1, int r2);
-void lasvm_kcache_swap_ii(lasvm_kcache_t *self, int i1, int i2);
-void lasvm_kcache_swap_ri(lasvm_kcache_t *self, int r1, int i2);
+void lasvm_kcache_swap_rr(lasvm_kcache_t *self, long r1, long r2);
+void lasvm_kcache_swap_ii(lasvm_kcache_t *self, long i1, long i2);
+void lasvm_kcache_swap_ri(lasvm_kcache_t *self, long r1, long i2);
 
 
 #ifdef __cplusplus__
